requirements.txt
psycopg2-binary==2.9.9
loguru==0.7.2
python-decouple==3.8
lxml==5.3.0
elasticsearch==8.15.1